+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Option to bypass the reporter entirely when there are no errors
+
 ## [1.4.0] - 2020-08-18
 ### Changed
 - [BREAKING CHANGE] Require Node.js engine v10 and higher